renamed fetch 3rd party script to install runtim dependencies
